Added tests for all views

+++ network/views.py
@@ -70,21 +70,22 @@ def register(request):
     else:
         return render(request, "network/register.html")
 
-
+@login_required(login_url="/login/")
 def post(request):
     if request.method == "POST":
         try:
             data = json.loads(request.body)
-            try:
-                content = data.get("content")
-            except:
-                return JsonResponse({"message": "Missing content parameter"}, status=400)
+        except:
+            return JsonResponse({"message": "Missing request body"}, status=400)
+        content = data.get("content")
+        if content is None:
+            return JsonResponse({"message": "Missing content parameter"}, status=400)
+        try:
             Post.objects.create(author=request.user, content=content)
-            return JsonResponse({"message": "Post created successfully"}, status=201)
         except:
             return HttpResponse(status=500)
-    else:
-        return JsonResponse({"message": "POST method required"}, status=400)
+        return JsonResponse({"message": "Post created successfully"}, status=201) 
+    return JsonResponse({"message": "POST method required"}, status=400)
 
 @login_required(login_url="/login/")
 def posts(request, id):
@@ -136,39 +137,47 @@ def profiles(request, id):
         if request.method == "GET":
             return JsonResponse(user.serialize(request.user), safe=False, status=200)
         if request.method == "PUT":
-            try:
-                data = json.loads(request.body)
-            except:
-                return JsonResponse({"message": "Missing request body"}, status=400)
-            follow = data.get("follow")
-            if follow is None:
-                return JsonResponse({"message": "Missing follow parameter"},status=400)
-            if follow == True:
-                if user.followers.filter(follower=request.user).count() > 0:
-                    return JsonResponse({"message": "You already follow this user"}, status=409)
-                Follow.objects.create(follower=request.user, followed=user)
-                return HttpResponse(status=204)        
-            if follow == False:
-                if user.followers.filter(follower=request.user).count() == 0:
-                    return JsonResponse({"message": "You already don't follow this user"}, status=409)
-                Follow.objects.get(follower=request.user, followed=user).delete()
-                return HttpResponse(status=204)
-            return JsonResponse({"message": "Invalid follow parameter"}, status=400)
+            if request.user.is_authenticated:
+                if request.user == user:
+                    return JsonResponse({"message": "You can't follow your own profile"}, status=400)
+                try:
+                    data = json.loads(request.body)
+                except:
+                    return JsonResponse({"message": "Missing request body"}, status=400)
+                follow = data.get("follow")
+                if follow is None:
+                    return JsonResponse({"message": "Missing follow parameter"},status=400)
+                if follow == True:
+                    if user.followers.filter(follower=request.user).count() > 0:
+                        return JsonResponse({"message": "You already follow this profile"}, status=409)
+                    Follow.objects.create(follower=request.user, followed=user)
+                    return HttpResponse(status=204)        
+                if follow == False:
+                    if user.followers.filter(follower=request.user).count() == 0:
+                        return JsonResponse({"message": "You already don't follow this profile"}, status=409)
+                    Follow.objects.get(follower=request.user, followed=user).delete()
+                    return HttpResponse(status=204)
+                return JsonResponse({"message": "Invalid follow parameter"}, status=400)
+            return HttpResponseRedirect(f"{reverse('login')}/?next={request.path}", status=302)
         return JsonResponse({"message": "Invalid request method"}, status=400)
     except:
         return HttpResponse(status=500)
     
 
 def all_posts(request):
-    posts = Post.objects.all().order_by("-date")
-    #return JsonResponse([post.toJson(request.user) for post in posts], safe=False)
-    #return JsonResponse({"post" : {"author": "author", "content": "content"}}, safe=False)
-    return JsonResponse([post.serialize(request.user) for post in posts], safe=False)
+    try:
+        posts = Post.objects.all().order_by("-date")
+        return JsonResponse([post.serialize(request.user) for post in posts], safe=False)
+    except:
+        return HttpResponse(status=500)
     
 @login_required(login_url="/login/")
 def following_posts(request):
-    posts = Post.objects.filter(author__in=User.objects.filter(followers__in=Follow.objects.filter(follower=request.user))).order_by("-date")
-    return JsonResponse([post.serialize(request.user) for post in posts], safe=False)
+    try:
+        posts = Post.objects.filter(author__in=User.objects.filter(followers__in=Follow.objects.filter(follower=request.user))).order_by("-date")
+        return JsonResponse([post.serialize(request.user) for post in posts], safe=False)
+    except:
+        return HttpResponse(status=500)
